+++ app/FeaturedPiece.php
@@ -6,15 +6,16 @@ use Illuminate\Database\Eloquent\Model;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 
-class Slide extends Model implements HasMedia
+class FeaturedPiece extends Model implements HasMedia
 {
     use InteractsWithMedia;
 
     protected $fillable = [
         'title',
-        'subtitle',
+        'excerpt',
         'url',
         'publish',
+        'type',
     ];
 
     public static $rules = [
@@ -22,6 +23,7 @@ class Slide extends Model implements HasMedia
         'url' => 'required',
         'publish' => 'boolean',
         'image' => 'image|dimensions:min_width=1200',
+        'type' => 'required|in:article,collection',
     ];
 
     public function scopePublished($query)