+++ resources/views/frontend/authors/form.blade.php
@@ -3,14 +3,17 @@
 <section class="filters">
     <div class="container content-section">
         <div class="row">
-            <div  class="col-md-4 col-xs-6 bottom-space">
+            <div  class="col-md-3 col-xs-6 bottom-space">
                 @formWidget($form['role'], ['attr' => ['class' => 'js-custom-select']])
             </div>
-            <div  class="col-md-4 col-xs-6 bottom-space">
+            <div  class="col-md-3 col-xs-6 bottom-space">
                 @formWidget($form['nationality'], ['attr' => ['class' => 'js-custom-select']])
             </div>
-            <div  class="col-md-4 col-xs-6 bottom-space">
+            <div  class="col-md-3 col-xs-6 bottom-space">
                 @formWidget($form['place'], ['attr' => ['class' => 'js-custom-select']])
+            </div>
+            <div  class="col-md-3 col-xs-6 bottom-space">
+                @formWidget($form['sex'], ['attr' => ['class' => 'js-custom-select']])
             </div>
         </div>
         <div class="row">