+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.41.0] - 2022-01-14
+### Added
+- filter by gender to authorities
+
+### Changed
+- inscription length in item_translations table
+- GMUHK work types
+- Slide to FeaturedPiece
+- no-image for alive/dead authors
+
+### Fixed
+- gender-specific no-image for authorities
+
 ## [2.40.1] - 2022-01-04
 ### Fixed
 - indexing multiple medium values